pyproject.toml
[tool.poetry]
name = "flo-ai"
version = "0.0.2.1"
description = "A easy way to create structured AI agents"
authors = ["vizsatiz <vishnu@rootfor.xyz>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
langchain = "0.2.11"
langgraph = "0.0.55"
ipython = "8.24.0"
httpx = "0.27.0"
langchain-community = "0.2.0"
langchain-experimental = "0.0.59"
pillow = "^10.3.0"
pydantic = "^2.9.2"


[tool.poetry.group.dev.dependencies]
langchain-mongodb = "^0.1.5"
langchain-chroma = "^0.1.1"
langchainhub = "^0.1.17"
pypdf = "^4.2.0"
ipykernel = "^6.29.5"
langchain-openai = "^0.1.8"
openai-ratelimiter = "^0.5"
langchain-anthropic = "^0.1.22"
db-sqlite3 = "^0.0.1"
peewee = "^3.17.6"
python-dotenv = "^1.0.1"
behave = "^1.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
